requirements.txt
# FastAPI framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Pydantic for data validation
pydantic>=2.5.0
python-dotenv>=1.0.0

# PyTorch and Transformers for MT5 model
torch>=2.1.0
transformers>=4.36.0
safetensors>=0.4.0
sentencepiece>=0.1.99

# CORS support
python-multipart>=0.0.6

# HTTP client for LLM API
httpx>=0.27.0

